fix(trustedcapability): make the framing-limit check 32-bit safe
codec.go used `int(^uint32(0))`, whose constant 4294967295 overflows a 32-bit
int, so `go build` fails on any 32-bit target (GOARCH=386/arm/mips...). This
only surfaced in consumers that cross-compile to 32-bit (e.g. tos-messenger).
Compare as uint64 instead — correct on all word sizes (on 32-bit a slice can
never exceed 2^31-1 anyway, so the check simply never trips there).

Verified: GOWORK=off GOARCH=386 go build ./... and the amd64 build both pass;
pkg/trustedcapability tests green.

+++ pkg/trustedcapability/codec.go
@@ -290,7 +290,7 @@ func ObjectDigest(object ProfileObjectV1) ([]byte, error) {
 	if err != nil {
 		return nil, err
 	}
-	if len(canonical) > int(^uint32(0)) {
+	if uint64(len(canonical)) > uint64(^uint32(0)) {
 		return nil, errors.New("canonical object exceeds framing limit")
 	}
 	hash := sha256.New()
